requirements.txt
transformers
sentence-transformers
torch
gradio
python-dotenv
